Move json functions to a separate module
